Update (deprecated) look-controls for WebVR v1.0
